Extract snapping fling behavior
